requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# OpenAI API
openai>=1.0.0
tiktoken>=0.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0